Moves earlier imgui customizations to their own .hpp and .cpp files
